# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Waylog API
  version: 0.2.0
  description: |
    HTTP API for the Waylog ingest server.

    The primary product path is schema-2.0 ingest plus the v2 read APIs used by
    the operator CLI and embedded dashboard: recent traces, error families,
    trace story, blast radius, event search, and direct trace/event lookup.

    Write endpoints require write-scope auth when auth is configured. Read
    endpoints require read-scope auth when read keys are configured.
servers:
  - url: /
...
paths:
  /v1/events:
    post:
      tags: [Ingest]
      operationId: ingestEvents
      summary: Ingest schema-2.0 WideEvents
      description: |
        Accepts one schema-2.0 WideEvent as JSON or a batch as NDJSON. Accepted
        events are written to the schema-2.0 WAL before projection into the v2
        read models. Per-event schema failures are returned inside the ingest
        envelope; infrastructure failures return 503.
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WideEvent'
          application/x-ndjson:
            schema:
              type: string
              description: Newline-delimited schema-2.0 WideEvent JSON objects.
      responses:
        '200':
          description: Ingest envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestEnvelope'
        '400':
          description: Malformed body, missing content type, unsupported encoding, or invalid cursor/body framing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestEnvelope'
        '401':
          description: Unauthorized
        '405':
          description: Method Not Allowed
        '413':
          description: Body or batch too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestEnvelope'
        '415':
          description: Unsupported content type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestEnvelope'
        '503':
          description: Durability or projection unavailable
...
  /v1/events/{event_id}:
    get:
      tags: [Events]
      operationId: getEvent
      summary: Get one schema-2.0 event
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventID'
      responses:
        '200':
          description: Event wrapper
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventGetResponse'
        '400':
          $ref: '#/components/responses/ReadBadRequest'
        '404':
          $ref: '#/components/responses/ReadNotFound'
  /v1/events/search:
    get:
      tags: [Events]
      operationId: searchEvents
      summary: Search schema-2.0 events
      description: Searches the v2 recent index with cursor pagination.
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Service'
        - $ref: '#/components/parameters/StatusCSV'
        - $ref: '#/components/parameters/ErrorCode'
        - $ref: '#/components/parameters/TraceIDQuery'
        - $ref: '#/components/parameters/Since'
        - $ref: '#/components/parameters/Until'
        - $ref: '#/components/parameters/Window'
        - $ref: '#/components/parameters/IncludeSuppressed'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Event search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventSearchResponse'
        '400':
          $ref: '#/components/responses/ReadBadRequest'
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key supplied via X-API-Key. Required for write or read operations when the corresponding server key is configured.
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token alternative to X-API-Key. Required for write or read operations when the corresponding server key is configured.
  parameters:
    EventID:
      name: event_id
      in: path
      required: true
      schema:
        type: string
...
    TraceIDQuery:
      name: trace_id
      in: query
      schema:
        type: string
    Service:
      name: service
      in: query
      schema:
        type: string
    StatusCSV:
      name: status
      in: query
      schema:
        type: string
      description: "Comma-separated status list: ok,error,timeout,partial,aborted,suppressed."
...
    ErrorCode:
      name: error_code
      in: query
      schema:
        type: string
    Since:
      name: since
      in: query
      schema:
        type: string
        format: date-time
    Until:
      name: until
      in: query
      schema:
        type: string
        format: date-time
    Window:
      name: window
      in: query
      schema:
        type: string
      description: Go duration string such as 15m or 1h.
    IncludeSuppressed:
      name: include_suppressed
      in: query
      schema:
        type: boolean
        default: false
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor from next_cursor.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
  responses:
    ReadBadRequest:
      description: Invalid query or request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReadError'
    ReadNotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReadError'
  schemas:
    IngestEnvelope:
      type: object
      required: [accepted, duplicate, rejected, deprecations]
      example:
        accepted: 1
        duplicate: 0
        rejected: []
        deprecations: {}
      properties:
        accepted:
          type: integer
        duplicate:
          type: integer
        rejected:
          type: array
          items:
            $ref: '#/components/schemas/RejectedEvent'
        deprecations:
          type: object
          additionalProperties:
            type: integer
    RejectedEvent:
      type: object
      required: [index, reason]
      properties:
        index:
          type: integer
        event_id:
          type: string
        reason:
          type: string
          enum:
            - invalid_json
            - schema_validation_failed
            - bridge_not_implemented
            - batch_oversize
            - body_oversize
            - unsupported_content_type
            - unsupported_content_encoding
            - invalid_body
            - durability_unavailable
        detail:
          type: string
    ReadError:
      type: object
      required: [error]
      example:
        error:
          code: invalid_query
          message: window must be a valid duration
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            detail:
              type: string
    WideEvent:
      type: object
      required: [schema_version, event_id, ts_start, ts_end, kind, service, env, trace_id, status]
      example:
        schema_version: "2.0"
        event_id: 11111111-1111-4111-8111-111111111111
        ts_start: "2026-05-04T10:00:00Z"
        ts_end: "2026-05-04T10:00:00.042Z"
        duration_ms: 42
        kind: http
        service: checkout
        env: demo
        version: local
        trace_id: 7f3a2b9c000000000000000000000001
        span_id: 0000000000000001
        status: error
        anchor:
          step: payment.charge
          error_code: PMT_502
          kind: downstream
        steps:
          - name: cart.validate
            start_ms: 0
            duration_ms: 1
            status: ok
          - name: db.load_cart
            start_ms: 1
            duration_ms: 4
            status: ok
            downstream:
              service: db
              endpoint: /cart/X1
              kind: http
          - name: inventory.reserve
            start_ms: 5
            duration_ms: 2
            status: ok
          - name: payment.charge
            start_ms: 7
            duration_ms: 35
            status: error
            downstream:
              service: payment
              endpoint: /charge
              kind: http
            error:
              code: PMT_502
              reason: upstream gateway 5xx
        logs:
          - ts_offset_ms: 5
            level: info
            msg: inventory reserved
          - ts_offset_ms: 41
            level: error
            msg: upstream gateway 5xx
        fields:
          demo:
            scenario: payment_502
            sku: X1
        errors:
          - code: PMT_502
            reason: upstream gateway 5xx
      properties:
        schema_version:
          type: string
          enum: ["2.0"]
        event_id:
          type: string
          format: uuid
        ts_start:
          type: string
          format: date-time
        ts_end:
          type: string
          format: date-time
        duration_ms:
          type: integer
          minimum: 0
        kind:
          type: string
          enum: [http]
        service:
          type: string
        env:
          type: string
        version:
          type: string
        trace_id:
          type: string
          pattern: '^[0-9a-f]{32}$'
        span_id:
          type: string
          pattern: '^([0-9a-f]{16})?$'
        parent_span_id:
          type: string
          pattern: '^([0-9a-f]{16})?$'
        status:
          type: string
          enum: [ok, error, timeout, partial, aborted, suppressed]
        anchor:
          $ref: '#/components/schemas/Anchor'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/EventStep'
        logs:
          type: array
          items:
            $ref: '#/components/schemas/EventLog'
        fields:
          type: object
          additionalProperties: true
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorRef'
    Anchor:
      type: object
      required: [step, error_code]
      properties:
        step:
          type: string
        error_code:
          type: string
        kind:
          type: string
    EventStep:
      type: object
      required: [name, start_ms, duration_ms, status]
      properties:
        name:
          type: string
        span_id:
          type: string
        start_ms:
          type: integer
          minimum: 0
        duration_ms:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [ok, error]
        downstream:
          $ref: '#/components/schemas/Downstream'
        error:
          type: object
          properties:
            code:
              type: string
            reason:
              type: string
            cause:
              type: string
    Downstream:
      type: object
      properties:
        service:
          type: string
        endpoint:
          type: string
        kind:
          type: string
    EventLog:
      type: object
      required: [ts_offset_ms, level, msg]
      properties:
        ts_offset_ms:
          type: integer
          minimum: 0
        level:
          type: string
          enum: [info, warn, error]
        msg:
          type: string
        fields:
          type: object
          additionalProperties: true
    ErrorRef:
      type: object
      required: [code]
      properties:
        code:
          type: string
        reason:
          type: string
    EventGetResponse:
      type: object
      required: [event]
      properties:
        event:
          $ref: '#/components/schemas/WideEvent'
    EventSearchResponse:
      type: object
      required: [events]
      example:
        events:
          - schema_version: "2.0"
            event_id: 11111111-1111-4111-8111-111111111111
            ts_start: "2026-05-04T10:00:00Z"
            ts_end: "2026-05-04T10:00:00.042Z"
            kind: http
            service: checkout
            env: demo
            trace_id: 7f3a2b9c000000000000000000000001
            status: error
        next_cursor: cur_01
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/WideEvent'
        next_cursor:
          type: string
          nullable: true
...